修改类型

+++ urban-common/src/main/java/com/unicom/urban/common/constant/SysConstants.java
@@ -6,7 +6,7 @@ public class SysConstants {
 
     public static final String DELETED_FILED = "deleted";
 
-    public static final int DELETED_VALUE = 0;
+    public static final long DELETED_VALUE = 0;
 
     public static final String WHERE_DELETE = DELETED_FILED + "=" + DELETED_VALUE;
